cli-go: Move ViolationsCount.ToString to ResultPrinter #361
